vent: Do not ignore /feeds and /package/feeds
For this fork we want to have a stable (static) feeds world using git
submodule references to specific tags of the relevant feed repos. This
avoids having to do a "./scripts/update -a" cycle in a freshly cloned
version of our fork.

The plan is to minimise any structural changes to this fork, to allow
easier pulls from the master if we do want to move forward. The only
differences should be some minor admin changes and our extra "vent"
packages.